+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this provider will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.15.0] - 2026-07-08
 
 ### Added
 
@@ -262,7 +262,7 @@ Initial public release.
   `DATAHUB_GMS_URL`/`DATAHUB_GMS_TOKEN` environment variables, or
   `~/.datahubenv` (DataHub CLI config).
 
-[Unreleased]: https://github.com/datahub-project/terraform-provider-datahub/compare/v0.14.0...HEAD
+[0.15.0]: https://github.com/datahub-project/terraform-provider-datahub/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/datahub-project/terraform-provider-datahub/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/datahub-project/terraform-provider-datahub/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/datahub-project/terraform-provider-datahub/compare/v0.11.0...v0.12.0